+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 3.1.4: 2024-12-05
 
 * Remove Optimole dashboard widget (T-23392)
+* Move Optimole under Settings (T-23392)
 
 ### 3.1.3: 2024-12-04